Move from header-only to header/source
C++20 module support is probably a year or two out from being supported
by GCC fully (or mostly) and then being supported by a release of the
offical arm-none-eabi-gcc compiler, so trying to keep everything in a
single source and suffering the build times due to this is not worth it
for now. A refactor will occur when module support is added.

This change will improve build times as it moves all of the heavier
headers into the source files and out of the headers, keeping the
headers as minimal as possible.