Migrate back to header only libraries for drivers
This is due to the fact that the conan does not make building from
source by a consumer basis easy to handle without a ton of work needed
in a the library consumers cmake file. To eliminate and simply this,
these libraries will be header only. The hope is that this can be
returned to h/cpp split files OR be converted into C++20 modules in
the future.