Split uart.hpp into interface/impl
Clean up the header file and class interface by removing most of the
implementation details. This way there is a clear interface and the
implementation is created below. The implementation can be copied and
pasted into a .cpp file at a later time when source buidling with conan
becomes feasible.

Finish untested implementation of UART driver.