Fix issues with UART interrupt receiver
UART RX Interrupt would result in a crash and this was due to a
get_uart<> buffer not being labeled as static resulting in it access
invalid memory. This also caused the data in the ring buffer to be
corrupted.

Additional fixes have been implemented to ensure proper opreation of the
UART port.