:bug: Fix FPU initialization for hard float demos
Use the __ARM_FP macro to call as the
`hal::cortex_m::try_initialize_floating_point_unit()` function is
actually unreliable due to the fact that vendor supplied floating
point IP blocks do not require bits to be set in that field. Meaning
that the function does not do as it is advertised.

NOTE: Deleted 2.0.0 version from package repo and moving version back to
2.0.0-alpha.1. This way we have an alpha release of 2.0.0 before
doing a full 2.0.0 release.

Resolves #174
Resolves #171