:technologist: Add debug info to all builds
CMake only adds the `-g` debug info flag for the "Debug" build type.
This makes sense for libraries deployed on a server or computer but not
an embedded system. The elf file is not what is executed but the .bin or
.hex file generated from it. Those generated files have all extra debug
info eliminated. Because of this, each build type can have debug
information without degrading performance or increasing code size. The
benefit is that users can use an on chip debugger and step into this
libraries APIs to inspect possible issues or hidden registers.