:bug: Fix i2c last byte read problem
lpc40::i2c::driver_transaction returns early during a read transaction.
The i2c interrupt continues to fire and fills the buffer after the
function has returned. This is problematic because it requires a delay
on the driver's part to manage this. The memory of the buffer may also
no longer be valid by the time the interrupt fires again.